fastapi/requirements.txt
# FastAPI フレームワーク
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.6.0

# データベース関連
sqlalchemy>=2.0.0
pymysql>=1.1.0
cryptography>=42.0.0
alembic>=1.13.0

# ユーティリティ
python-dotenv>=1.0.0
python-multipart>=0.0.9

# CORS対応